+++ src/App.tsx
@@ -1,10 +1,8 @@
-
 import Routes from 'routes/Routes';
 
 function App() {
   return (
     <div className='App'>
-
       <Routes />
     </div>
   );
