--- src/assets/index.ts
@@ -1,1 +1,0 @@
-export {};
