--- src/index.css
@@ -1,12 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@font-face {
-  font-family: 'Neo';
-  src: url(./fonts/neodgm.ttf);
-}
-
-:root {
-  font-family: 'Neo';
-}
