+++ src/index.tsx
@@ -1,4 +1,3 @@
-
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
@@ -17,5 +16,4 @@ root.render(
       <App />
     </BrowserRouter>
   </QueryClientProvider>,
-
 );
