--- src/routes/index.tsx
@@ -1,18 +1,0 @@
-import { Navigate, Route, Routes as ReactRouters } from 'react-router-dom';
-import Layout from '../components/Layout';
-import MainPage from '../pages/mainpage/MainPage';
-import PostPage from '../pages/PostPage';
-
-const Routes = () => {
-  return (
-    <ReactRouters>
-      <Route element={<Layout />}>
-        <Route path='/' element={<MainPage />} />
-        <Route path='/post' element={<PostPage />} />
-        <Route path='*' element={<Navigate replace to='/' />} />
-      </Route>
-    </ReactRouters>
-  );
-};
-
-export default Routes;
